+++ ast.go
@@ -30,11 +30,12 @@ func (Comment) isStatement() {}
 
 // Block represents a container directive, e.g. `<Directory /var/www> ... </Directory>`.
 type Block struct {
-	Name     string
-	Args     []string
-	Children []Statement
-	Pos      Position
-	EndPos   Position
+	Name       string
+	Args       []string
+	Children   []Statement
+	Pos        Position
+	EndPos     Position
+	EndComment string
 }
 
 func (Block) isStatement() {}